Bugfixes related to 4to8 and gameobject structs. Is now printing currently loaded gameobjects with actor names.

+++ include/cpp/cpp_handler.h
@@ -3,7 +3,7 @@
 #include <filesystem>
 
 void update_loaded_gameobjects();
-bool api_init_cpp(uintptr_t base_address, const std::filesystem::path &path);
+bool api_init_cpp(uint64_t base_address, const std::filesystem::path &path);
 void load_mod_setup_cpp();
 void load_mods_cpp();
 void unload_mods_cpp();
